update: Move command highlighting to acdef.js script
- Since the first command (main command) gets removed from each
command chain, highlighting the command before removing the main
command now introduces the problem that the command chain is wrapped
with ANSI color codes. This causes the RegExp to not catch the
main command. One of two things can be done:

1) Highlight chain after removing the main command.
2) Or update RegExp to be aware of the ANSI code.

This commit opts for option 1 as option 2 will also require an
update to the RegExp in the case that the ANSI color/styles change
later on.