feature: Add pre-hook to manipulate ACDEF file contents on the fly
- This was primarily added for yarn. Yarn provides a shortcut way
to run a repo's scripts. Simply by doing '$yarn SCRIPT-NAME' will
run the provided script. This posed issues as getting scripts needs
to be done on the fly and yarn.acdef is a statically generated file.
Therefore, the contents of yarn.acdef don't have the script names.
Modifying the contents of yarn.acmap can now be done via a pre-hook
script.
- In essence, nodecliac will look for the hook script and if found
it will run it. Here the auto-completion developer has the ability
to modify the acdef file contents.
- Note: The actual contents of the acdef file do not get modified.
These modifications are done in memory. The original file does not
get modified.