fix: Correct 'nodecliac bin -' tab completion (Nim)
- Note: A trailing '--' is used to denote 'no flags' in an
.acdef row. Therefore, when a command chain has no flags
this must be accounted for. As the flags array still contains
the empty '--'. Or else when a [Tab][Tab] is performed,
say 'nodecliac bin -' for example, the '-' will be completed
to '--'. Which is what we do not want. Again, this is the
case as there is technically one flag in the flags array
(i.e. @["--"]).